+++ src/main/java/com/kalsym/order/service/model/CustomerVoucher.java
@@ -49,6 +49,7 @@ public class CustomerVoucher implements Serializable {
     private String customerId;    
     private String voucherId;    
     private Boolean isUsed;
+    private String guestEmail;
     
     @Temporal(javax.persistence.TemporalType.TIMESTAMP)
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")